requirements_dev.txt
# development
black
coverage[toml]
flake8
flit==3.7.1
ipdb
isort
mypy
pre-commit-hooks
pytest

# docs
# it is important to freeze the version because the option -W is used for sphinx-build
sphinx==4.2.0
sphinx-copybutton
sphinx-material
sphinxcontrib-spelling

# examples
torchvision==0.8.2

# other
jupyterlab
